Moving supporting models to single directory
It may be helpful to have full model context for writing specs.
