--- spec/support/models/newspaper_issue.rb
@@ -1,9 +1,0 @@
-# Newspaper Issue
-class NewspaperIssue < ActiveFedora::Base
-  # WorkBehavior mixes in minimal ::Hyrax::CoreMetadata fields of
-  # depositor, title, date_uploaded, and date_modified.
-  # https://samvera.github.io/customize-metadata-model.html#core-metadata
-  include ::Hyrax::WorkBehavior
-  # BasicMetadata must be included last
-  include ::Hyrax::BasicMetadata
-end
